move worldly pattern renderers out of PatternTextureManager
